# controllers/user.go
package controllers

import (
	"github.com/gin-gonic/gin"
	"gin-frame/models"
	. "gin-frame/utils"
	"net/http"
		)

func UserSetup(v1 *gin.RouterGroup) {
	r := v1.Group("/user")

	r.GET("login", login)
	r.POST("logout", logout)

}

// @Tags user
// @Summary 登录
// @Accept json
// @Produce  json
// @Param name query string true "姓名备注"
// @Router /user/login [get]
func login(c *gin.Context) {
	user := models.Users{Name: "cyh", Age: 18, Email: "abcdefg@qq.com"}
	err := Orm.Create(&user).Error
	if err != nil {
		Log.Errorln(err)
	}

	var searchUser models.Users
	err = Orm.Where("name = ?", "cyh").First(&searchUser).Error
	if err != nil {
		Log.Errorln(err)
	} else {
		Log.WithField("user", searchUser).Info("test")
	}

	searchUser.Age = 24
	err = Orm.Save(&searchUser).Error
	if err != nil {
		Log.Errorln(err)
	}

	c.JSON(http.StatusOK, gin.H{
		"user": searchUser,
	})
}

type Info struct {
	Name string `json:"name1"`
	Age  string
	Pig  int `json:"pig"`
}

// @Tags user
// @Summary 登出
// @Accept json
// @Produce  json
// @Param name  body 		controllers.Info	true 	"test"
// @Param age  	formData 	string				true 	"test"
// @Param fff  	query 	string				true 	"test"
// @Router /user/logout [post]
func logout(c *gin.Context) {
	age := c.PostForm("age")
	Log.Infoln("form age", age)

	Log.Infoln("get age", c.GetString("age"))

	Log.Infoln("get name1", c.GetString("name1"))


	c.JSON(http.StatusOK, gin.H{
		"user": 123,
	})
}